# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Plan {
  id: number;
  name: string;
  /** sd_tv | hd_tv | internet_only */
  category: string;
  /** Speed in Mbps */
  speed: number;
  /** Monthly price in INR */
  price: number;
  features: string[];
  isPopular: boolean;
  createdAt: string;
}

export interface Connection {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  whatsapp?: string | null;
  address: string;
  village: string;
  pinCode: string;
  /** @nullable */
  planId?: number | null;
  /** home | business */
  connectionType: string;
  /** @nullable */
  installationDate?: string | null;
  /** pending | approved | installed | rejected */
  status: string;
  createdAt: string;
}

export interface ConnectionInput {
  fullName: string;
  mobile: string;
  whatsapp?: string;
  address: string;
  village: string;
  pinCode: string;
  planId?: number;
  connectionType: string;
  installationDate?: string;
}

export interface ConnectionUpdate {
  status?: string;
}

export interface OtpRequest {
  mobile: string;
}

export interface OtpResponse {
  success: boolean;
  message: string;
}

export interface OtpVerify {
  mobile: string;
  otp: string;
}

export interface Customer {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  address: string;
  village: string;
  pinCode: string;
  /** @nullable */
  planId?: number | null;
  /** active | inactive | suspended */
  status: string;
  dataUsedGB: number;
  dataLimitGB: number;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  customer: Customer;
}

export interface Complaint {
  id: number;
  /** @nullable */
  customerId?: number | null;
  name: string;
  mobile: string;
  subject: string;
  description: string;
  /** open | in_progress | resolved | closed */
  status: string;
  /** low | medium | high */
  priority: string;
  createdAt: string;
}

export interface ComplaintInput {
  customerId?: number;
  name: string;
  mobile: string;
  subject: string;
  description: string;
  priority?: string;
}

export interface ComplaintUpdate {
  status?: string;
  priority?: string;
}

export interface AdminStats {
  totalCustomers: number;
  activeConnections: number;
  pendingRequests: number;
  openComplaints: number;
  monthlyRevenue: number;
  newConnectionsThisMonth: number;
}

export type ListConnectionsParams = {
status?: string;
};

export type ListCustomersParams = {
search?: string;
};

export type ListComplaintsParams = {
status?: string;
};