# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalCustomers: number;
  activeConnections: number;
  pendingRequests: number;
  openComplaints: number;
  monthlyRevenue: number;
  newConnectionsThisMonth: number;
}