# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface Complaint {
  id: number;
  /** @nullable */
  customerId?: number | null;
  name: string;
  mobile: string;
  subject: string;
  description: string;
  /** open | in_progress | resolved | closed */
  status: string;
  /** low | medium | high */
  priority: string;
  createdAt: string;
}