# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintInput {
  customerId?: number;
  name: string;
  mobile: string;
  subject: string;
  description: string;
  priority?: string;
}