# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface Connection {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  whatsapp?: string | null;
  address: string;
  village: string;
  pinCode: string;
  /** @nullable */
  planId?: number | null;
  /** home | business */
  connectionType: string;
  /** @nullable */
  installationDate?: string | null;
  /** pending | approved | installed | rejected */
  status: string;
  createdAt: string;
}