# lib/api-zod/src/generated/types/connectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionInput {
  fullName: string;
  mobile: string;
  whatsapp?: string;
  address: string;
  village: string;
  pinCode: string;
  planId?: number;
  connectionType: string;
  installationDate?: string;
}