# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  fullName: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  address: string;
  village: string;
  pinCode: string;
  /** @nullable */
  planId?: number | null;
  /** active | inactive | suspended */
  status: string;
  dataUsedGB: number;
  dataLimitGB: number;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}