# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rudra Fiber Net API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  /** sd_tv | hd_tv | internet_only */
  category: string;
  /** Speed in Mbps */
  speed: number;
  /** Monthly price in INR */
  price: number;
  features: string[];
  isPopular: boolean;
  createdAt: string;
}